services/transformers/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
lxml==5.3.0
xmltodict==0.14.2
python-liquid==1.12.0
